test: prove Code Mode approval clock rollback
